🚚: Reorganised existing sources
Moved sources in an /src tree to ease next refactor steps.
Current unit test and program outputs have been checked.
