+++ src/model/drug/drug.js
@@ -1,0 +1,7 @@
+export class Drug {
+  constructor(name, expiresIn, benefit) {
+    this.name = name;
+    this.expiresIn = expiresIn;
+    this.benefit = benefit;
+  }
+}
