+++ src/model/pharmacy/pharmacy.js
@@ -1,15 +1,8 @@
-export class Drug {
-  constructor(name, expiresIn, benefit) {
-    this.name = name;
-    this.expiresIn = expiresIn;
-    this.benefit = benefit;
-  }
-}
-
 export class Pharmacy {
   constructor(drugs = []) {
     this.drugs = drugs;
   }
+
   updateBenefitValue() {
     for (var i = 0; i < this.drugs.length; i++) {
       if (
